requirements.txt
flask
flask-sqalchemy
flask-migrate
flask-login
flask-bcrypt
flask-wtf
python-dotenv